Fix bugs on the UserBoardFragment due to incorrect visibility settings.
Fix bugs when the menu icon and board creation button were displayed if
the current workspace was null.
Fix a bug when the tip text were not displayed if the current workspace
was null.
Fix a bug when the list of user workspaces was not updated after
creating a workspace.